src/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
crispy-bootstrap5==0.7
distlib==0.3.6
Django==4.1.5
django-crispy-forms==1.14.0
filelock==3.9.0
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2==2.9.5
sqlparse==0.4.3
virtualenv==20.17.1
virtualenv-clone==0.5.7
